+++ lumicks/pylake/file.py
@@ -2,8 +2,8 @@ import h5py
 import numpy as np
 from typing import Dict
 
-from .channel import make_continuous_channel, make_timeseries_channel, Slice, Timeseries
-from .detail.mixin import Force, DownsampledFD, PhotonCounts
+from .channel import Slice, Continuous, TimeSeries, TimeTags, channel_class
+from .detail.mixin import Force, DownsampledFD, PhotonCounts, PhotonTimeTags
 from .fdcurve import FDCurve
 from .group import Group
 from .kymo import Kymo
@@ -13,7 +13,7 @@ from .scan import Scan
 __all__ = ["File"]
 
 
-class File(Group, Force, DownsampledFD, PhotonCounts):
+class File(Group, Force, DownsampledFD, PhotonCounts, PhotonTimeTags):
     """A convenient HDF5 file wrapper for reading data exported from Bluelake
 
     Parameters
@@ -31,14 +31,28 @@ class File(Group, Force, DownsampledFD, PhotonCounts):
         file.force1x.plot()
         file.kymos["name"].plot()
     """
+
+    SUPPORTED_FILE_FORMAT_VERSIONS = [1, 2]
+
     def __init__(self, filename):
         super().__init__(h5py.File(filename, 'r'))
+        self._check_file_format()
+
+    def _check_file_format(self):
+        if "Bluelake version" not in self.h5.attrs:
+            raise Exception("Invalid HDF5 file: no Bluelake version tag found")
+        if "File format version" not in self.h5.attrs:
+            raise Exception("Invalid HDF5 file: no file format version tag found")
+        ff_version = int(self.h5.attrs["File format version"])
+        if ff_version not in File.SUPPORTED_FILE_FORMAT_VERSIONS:
+            raise Exception(f"Unsupported Bluelake file format version {ff_version}")
 
     @classmethod
     def from_h5py(cls, h5py_file):
         """Directly load an existing `h5py.File`"""
         new_file = cls.__new__(cls)
-        super(cls, new_file).__init__(h5py_file)
+        new_file.h5 = h5py_file
+        new_file._check_file_format()
         return new_file
 
     @property
@@ -105,13 +119,13 @@ class File(Group, Force, DownsampledFD, PhotonCounts):
         return print_attributes(self.h5) + "\n" + print_group(self.h5)
 
     def _get_force(self, n, xy):
-        return make_continuous_channel(self.h5["Force HF"][f"Force {n}{xy}"], "Force (pN)")
+        return Continuous.from_dataset(self.h5["Force HF"][f"Force {n}{xy}"], "Force (pN)")
 
     def _get_downsampled_force(self, n, xy):
         group = self.h5["Force LF"]
 
         def make(channel):
-            return make_timeseries_channel(group[channel], "Force (pN)")
+            return TimeSeries.from_dataset(group[channel], "Force (pN)")
 
         if xy:  # An x or y component of the downsampled force is easy
             return make(f"Force {n}{xy}")
@@ -125,15 +139,18 @@ class File(Group, Force, DownsampledFD, PhotonCounts):
         # If it's completely missing, we can reconstruct it from the x and y components
         fx = make(f"Force {n}x")
         fy = make(f"Force {n}y")
-        return Slice(Timeseries(np.sqrt(fx.data**2 + fy.data**2), fx.timestamps),
+        return Slice(TimeSeries(np.sqrt(fx.data**2 + fy.data**2), fx.timestamps),
                      labels={"title": f"Force LF/Force {n}", "y": "Force (pN)"})
 
     def _get_distance(self, n):
-        return make_timeseries_channel(self.h5["Distance"][f"Distance {n}"],
+        return TimeSeries.from_dataset(self.h5["Distance"][f"Distance {n}"],
                                        r"Distance ($\mu$m)")
 
     def _get_photon_count(self, name):
-        return make_continuous_channel(self.h5["Photon count"][name], "Photon count")
+        return Continuous.from_dataset(self.h5["Photon count"][name], "Photon count")
+
+    def _get_photon_time_tags(self, name):
+        return TimeTags.from_dataset(self.h5["Photon Time Tags"][name], "Photon time tags")
 
     @property
     def kymos(self) -> Dict[str, Kymo]: