+++ lumicks/pylake/tests/test_channels.py
@@ -5,13 +5,18 @@ from lumicks.pylake import channel
 
 def test_slice_properties():
     size = 5
-    s = channel.Slice(channel.Timeseries(np.random.rand(size), np.random.rand(size)))
+    s = channel.Slice(channel.TimeSeries(np.random.rand(size), np.random.rand(size)))
     assert len(s) == size
     assert s.sample_rate is None
 
     s = channel.Slice(channel.Continuous(np.random.rand(size), start=0, dt=1))
     assert len(s) == size
     assert s.sample_rate == 1e9
+
+    size = 10
+    s = channel.Slice(channel.TimeTags(np.arange(0, size, dtype=np.int64)))
+    assert len(s) == size
+    assert s.sample_rate is None
 
     s = channel.empty_slice
     assert len(s) == 0
@@ -22,13 +27,13 @@ def test_labels():
     """Slicing must preserve labels"""
     size = 5
     labels = {"x": "distance", "y": "force"}
-    s = channel.Slice(channel.Timeseries(np.random.rand(size), np.random.rand(size)), labels)
+    s = channel.Slice(channel.TimeSeries(np.random.rand(size), np.random.rand(size)), labels)
     assert s.labels == labels
     assert s[:].labels == labels
     assert s[:0].labels == labels
     assert s[:10].labels == labels
 
-    s = channel.Slice(channel.Timeseries([], []), labels)
+    s = channel.Slice(channel.TimeSeries([], []), labels)
     assert len(s) == 0
     assert s.labels == labels
     assert s[:].labels == labels
@@ -42,7 +47,7 @@ def test_empty_slice():
 
 def test_timeseries_indexing():
     """The default integer indices are in timestamps (ns)"""
-    s = channel.Slice(channel.Timeseries([14, 15, 16, 17], [4, 5, 6, 7]))
+    s = channel.Slice(channel.TimeSeries([14, 15, 16, 17], [4, 5, 6, 7]))
 
     np.testing.assert_equal(s[0:5].data, [14])
     np.testing.assert_equal(s[0:5].timestamps, [4])
@@ -60,7 +65,7 @@ def test_timeseries_indexing():
         assert s[1:2:3]
     assert str(exc.value) == "Slice steps are not supported"
 
-    s = channel.Slice(channel.Timeseries([], []))
+    s = channel.Slice(channel.TimeSeries([], []))
     assert len(s[1:2].data) == 0
     assert len(s[1:2].timestamps) == 0
 
@@ -93,14 +98,36 @@ def test_continuous_idexing():
         assert s[1:2:3]
     assert str(exc.value) == "Slice steps are not supported"
 
-    s = channel.Slice(channel.Timeseries([], []))
+    s = channel.Slice(channel.TimeSeries([], []))
     assert len(s[1:2].data) == 0
     assert len(s[1:2].timestamps) == 0
 
 
+def test_timetags_indexing():
+    s = channel.Slice(channel.TimeTags([10, 20, 30, 40, 50, 60]))
+    np.testing.assert_equal(s[0:100].data, [10, 20, 30, 40, 50, 60])
+    np.testing.assert_equal(s[10:100].data, [10, 20, 30, 40, 50, 60])
+    np.testing.assert_equal(s[15:100].data, [20, 30, 40, 50, 60])
+    np.testing.assert_equal(s[10:60].data, [10, 20, 30, 40, 50])
+    np.testing.assert_equal(s[10:55].data, [10, 20, 30, 40, 50])
+    np.testing.assert_equal(s[11:50].data, [20, 30, 40])
+    np.testing.assert_equal(s[20:].data, [20, 30, 40, 50, 60])
+    np.testing.assert_equal(s[:50].data, [10, 20, 30, 40])
+
+    with pytest.raises(IndexError) as exc:
+        assert s[1]
+    assert str(exc.value) == "Scalar indexing is not supported, only slicing"
+    with pytest.raises(IndexError) as exc:
+        assert s[1:2:3]
+    assert str(exc.value) == "Slice steps are not supported"
+
+    s = channel.Slice(channel.TimeTags([]))
+    assert len(s[10:30].data) == 0
+
+
 def test_time_indexing():
     """String time-based indexing"""
-    s = channel.Slice(channel.Timeseries([1, 2, 3, 4, 5], [1400, 2500, 16e6, 34e9, 122 * 1e9]))
+    s = channel.Slice(channel.TimeSeries([1, 2, 3, 4, 5], [1400, 2500, 16e6, 34e9, 122 * 1e9]))
     # --> in time indices: ['0ns', '1100ns', '15.9986ms', '33.99s', '2m 2s']
 
     def assert_equal(actual, expected):
@@ -138,18 +165,25 @@ def test_time_indexing():
 
 
 def test_inspections(h5_file):
-    assert channel.is_continuous_channel(h5_file["Force HF"]["Force 1x"]) is True
-    assert channel.is_continuous_channel(h5_file["Force LF"]["Force 1x"]) is False
+    assert channel.channel_class(h5_file["Force HF"]["Force 1x"]) == channel.Continuous
+    assert channel.channel_class(h5_file["Force LF"]["Force 1x"]) == channel.TimeSeries
+    if "Photon Time Tags" in h5_file:
+        assert channel.channel_class(h5_file["Photon Time Tags"]["Red"]) == channel.TimeTags
 
 
 def test_channel(h5_file):
-    force = channel.make_continuous_channel(h5_file["Force HF"]["Force 1x"])
+    force = channel.Continuous.from_dataset(h5_file["Force HF"]["Force 1x"])
     assert np.allclose(force.data, [0, 1, 2, 3, 4])
     assert np.allclose(force.timestamps, [1, 11, 21, 31, 41])
 
-    downsampled = channel.make_timeseries_channel(h5_file["Force LF"]["Force 1x"])
+    downsampled = channel.TimeSeries.from_dataset(h5_file["Force LF"]["Force 1x"])
     assert np.allclose(downsampled.data, [1.1, 2.1])
     assert np.allclose(downsampled.timestamps, [1, 2])
+
+    if "Photon Time Tags" in h5_file:
+        timetags = channel.TimeTags.from_dataset(h5_file["Photon Time Tags"]["Red"])
+        assert np.all(np.equal(timetags.data, [10, 20, 30, 40, 50, 60, 70, 80, 90]))
+        assert np.all(np.equal(timetags.timestamps, [10, 20, 30, 40, 50, 60, 70, 80, 90]))
 
 
 def test_downsampling():