+++ changelog.md
@@ -3,6 +3,7 @@
 ## v0.3.0 | In development
 
 * TIFF scan exports now default to using 32bit floats
+* Support for photon time tag data has been added (Bluelake version 1.5.0 and higher, HDF5 file format version 2)
 
 ## v0.2.0 | 2018-07-27